% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_packages.R
\name{analyze_package_status}
\alias{analyze_package_status}
\title{Analyze package status}
\usage{
analyze_package_status(package_defs)
}
\arguments{
\item{package_defs}{definitions}
}
\value{
analysis
}
\description{
Analyze package status
}
\keyword{internal}
