% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_helper.R
\name{build_cloud_url}
\alias{build_cloud_url}
\title{Build cloud function URL}
\usage{
build_cloud_url(input, region = "ap-shanghai")
}
\arguments{
\item{input}{user input}

\item{region}{cloud function region}
}
\value{
complete URL
}
\description{
Build cloud function URL
}
\keyword{internal}
