% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_packages.R
\name{check_packages}
\alias{check_packages}
\title{R Package Environment Check and Dependency Installation}
\usage{
check_packages(interactive = TRUE, install_missing = TRUE)
}
\arguments{
\item{interactive}{interactive mode}

\item{install_missing}{auto install}
}
\value{
Returns list
}
\description{
R Package Environment Check and Dependency Installation
}
