% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_packages.R
\name{check_special_packages}
\alias{check_special_packages}
\title{Check special packages that cannot be installed from CRAN}
\usage{
check_special_packages(special_packages)
}
\arguments{
\item{special_packages}{list of special packages}
}
\description{
Check special packages that cannot be installed from CRAN
}
\keyword{internal}
