% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_package.R
\name{cleanup_temp_files}
\alias{cleanup_temp_files}
\title{Clean up temporary files}
\usage{
cleanup_temp_files(temp_file)
}
\arguments{
\item{temp_file}{path}
}
\description{
Clean up temporary files
}
\keyword{internal}
