% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_helper.R
\name{detect_input_type}
\alias{detect_input_type}
\title{Detect input type}
\usage{
detect_input_type(input)
}
\arguments{
\item{input}{user input string}
}
\value{
string
}
\description{
Detect input type
}
\keyword{internal}
