% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_package.R
\name{detect_system_environment}
\alias{detect_system_environment}
\title{Detect system environment}
\usage{
detect_system_environment()
}
\value{
list
}
\description{
Detect system environment
}
\keyword{internal}
