% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_code_internal.R
\name{.mc_check_hardware_warning}
\alias{.mc_check_hardware_warning}
\title{Check hardware info and stop if insufficient}
\usage{
.mc_check_hardware_warning(os_type, hw_info, quiet = FALSE)
}
\description{
Check hardware info and stop if insufficient
}
\keyword{internal}
