% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_code_internal.R
\name{.mc_compute_hash}
\alias{.mc_compute_hash}
\title{Compute hash from system information}
\usage{
.mc_compute_hash(os_type, computer_name, user_name, persistent_uuid, hw_info)
}
\description{
Compute hash from system information
}
\keyword{internal}
