% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_code_internal.R
\name{.mc_format_code}
\alias{.mc_format_code}
\title{Format hash as machine code with OS-specific prefix}
\usage{
.mc_format_code(os_type, hash)
}
\description{
Format hash as machine code with OS-specific prefix
}
\keyword{internal}
