% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_code_internal.R
\name{.mc_generate}
\alias{.mc_generate}
\title{Generate machine code - Enhanced with UUID and system identifiers}
\usage{
.mc_generate()
}
\description{
Generate machine code - Enhanced with UUID and system identifiers
}
\keyword{internal}
