% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_code_internal.R
\name{.mc_get_hardware_id}
\alias{.mc_get_hardware_id}
\title{Get hardware identifier based on OS}
\usage{
.mc_get_hardware_id(os_type)
}
\description{
Get hardware identifier based on OS
}
\keyword{internal}
