% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_code_internal.R
\name{.mc_get_hardware_info}
\alias{.mc_get_hardware_info}
\title{Get hardware information based on OS}
\usage{
.mc_get_hardware_info()
}
\description{
Get hardware information based on OS
}
\keyword{internal}
