% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_code_internal.R
\name{.mc_get_linux_hardware}
\alias{.mc_get_linux_hardware}
\title{Get Linux hardware information (returns vector like old version)}
\usage{
.mc_get_linux_hardware()
}
\description{
Get Linux hardware information (returns vector like old version)
}
\keyword{internal}
