% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_code_internal.R
\name{.mc_get_quiet}
\alias{.mc_get_quiet}
\title{Internal function to get machine code quietly (for install_package use)}
\usage{
.mc_get_quiet()
}
\description{
Internal function to get machine code quietly (for install_package use)
}
\keyword{internal}
