% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_code_internal.R
\name{.mc_get_windows_id}
\alias{.mc_get_windows_id}
\title{Get Windows MachineGuid from registry}
\usage{
.mc_get_windows_id()
}
\description{
Get Windows MachineGuid from registry
}
\keyword{internal}
