% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_code_internal.R
\name{.mc_handler}
\alias{.mc_handler}
\title{MC Handler - Main entry point for MC functionality}
\usage{
.mc_handler()
}
\description{
MC Handler - Main entry point for MC functionality
}
\keyword{internal}
