% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_package.R
\name{download_package_file}
\alias{download_package_file}
\title{Download package file}
\usage{
download_package_file(download_url, os_type)
}
\arguments{
\item{download_url}{URL}

\item{os_type}{type}
}
\value{
path
}
\description{
Download package file
}
\keyword{internal}
