% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_packages.R
\name{generate_environment_report}
\alias{generate_environment_report}
\title{Generate environment report}
\usage{
generate_environment_report(sys_config)
}
\arguments{
\item{sys_config}{system configuration}
}
\value{
environment information
}
\description{
Generate environment report
}
\keyword{internal}
