% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_packages.R
\name{generate_final_report}
\alias{generate_final_report}
\title{Generate final report}
\usage{
generate_final_report(package_defs, pkg_analysis)
}
\arguments{
\item{package_defs}{definitions}

\item{pkg_analysis}{analysis results}
}
\value{
final results
}
\description{
Generate final report
}
\keyword{internal}
