% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_packages_config.R
\name{get_package_definitions}
\alias{get_package_definitions}
\title{Get package definitions}
\usage{
get_package_definitions()
}
\value{
list
}
\description{
Get package definitions
}
\keyword{internal}
