% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_packages.R
\name{handle_library_path_configuration}
\alias{handle_library_path_configuration}
\title{Handle library path configuration}
\usage{
handle_library_path_configuration(interactive, sys_config)
}
\arguments{
\item{interactive}{interactive mode}

\item{sys_config}{system configuration}
}
\value{
path
}
\description{
Handle library path configuration
}
\keyword{internal}
