% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_packages_config.R
\name{initialize_system_config}
\alias{initialize_system_config}
\title{Initialize system configuration}
\usage{
initialize_system_config()
}
\value{
list
}
\description{
Initialize system configuration
}
\keyword{internal}
