% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_package.R
\name{install_and_verify_package}
\alias{install_and_verify_package}
\title{Install and verify package}
\usage{
install_and_verify_package(temp_file, package_name, lib_path, extract_func)
}
\arguments{
\item{temp_file}{path}

\item{package_name}{name of the package}

\item{lib_path}{path}

\item{extract_func}{extraction function (deprecated, kept for compatibility)}
}
\value{
TRUE
}
\description{
Install and verify package
}
\keyword{internal}
