% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_packages.R
\name{install_essential_packages}
\alias{install_essential_packages}
\title{Install}
\usage{
install_essential_packages(pkg_analysis, package_defs, interactive)
}
\arguments{
\item{pkg_analysis}{analysis results}

\item{package_defs}{definitions}

\item{interactive}{interactive mode}
}
\value{
installation results
}
\description{
Install
}
\keyword{internal}
