% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_code.R
\name{machine_code}
\alias{machine_code}
\title{Generate machine code}
\usage{
machine_code(quiet = FALSE)
}
\arguments{
\item{quiet}{suppress output}
}
\value{
formatted machine code
}
\description{
Generate unique machine identifier
}
\keyword{internal}
