% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_packages_ui.R
\name{persist_library_path}
\alias{persist_library_path}
\title{Persist library path configuration}
\usage{
persist_library_path(lib_path)
}
\arguments{
\item{lib_path}{Library path to persist}
}
\value{
list with success status
}
\description{
Persist library path configuration
}
\keyword{internal}
