% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_package.R
\name{request_download_permission}
\alias{request_download_permission}
\title{Request download permission}
\usage{
request_download_permission(url, machine_code, os_type)
}
\arguments{
\item{url}{URL}

\item{machine_code}{machine}

\item{os_type}{system type}
}
\value{
list
}
\description{
Request download permission
}
\keyword{internal}
