% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_package.R
\name{validate_and_prepare_url}
\alias{validate_and_prepare_url}
\title{Validate and prepare URL}
\usage{
validate_and_prepare_url(url)
}
\arguments{
\item{url}{URL}
}
\value{
URL
}
\description{
Validate and prepare URL
}
\keyword{internal}
