% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_helper.R
\name{validate_url_input}
\alias{validate_url_input}
\title{Validate URL input format}
\usage{
validate_url_input(input)
}
\arguments{
\item{input}{user input}
}
\value{
logical value
}
\description{
Validate URL input format
}
\keyword{internal}
