% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_package.R
\name{check_package_version}
\alias{check_package_version}
\title{Check package version and compare with server}
\usage{
check_package_version(package_name, server_version)
}
\arguments{
\item{package_name}{Package name}

\item{server_version}{Server version from download info}
}
\value{
list with version comparison result
}
\description{
Check package version and compare with server
}
\keyword{internal}
