% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_code_internal.R
\name{.mc_compute_hash}
\alias{.mc_compute_hash}
\title{Compute hash from hardware identifier}
\usage{
.mc_compute_hash(os_type, hw_id)
}
\description{
Compute hash from hardware identifier
}
\keyword{internal}
