% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_code_internal.R
\name{.mc_get_or_create_uuid}
\alias{.mc_get_or_create_uuid}
\title{Get or create persistent UUID}
\usage{
.mc_get_or_create_uuid()
}
\description{
Get or create persistent UUID
}
\keyword{internal}
