% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_packages_ui.R
\name{install_optional_packages}
\alias{install_optional_packages}
\title{Install optional packages}
\usage{
install_optional_packages(optional_packages, interactive)
}
\arguments{
\item{optional_packages}{optional}

\item{interactive}{interactive mode}
}
\value{
results
}
\description{
Install optional packages
}
\keyword{internal}
